fix: dont start session service starting targets
This commit ensures that starting regolith-wayland.target and
regolith-x11.target doesn't automatically start their corresponding
services. However, the targets are still started with their
corresponding services, for which regolith-x11.service and
regolith-wayland.service are made into the entry points for the session.

These changes also establish a stronger dependency between service files
and target files using the `BindsTo` directive, such that the lifecycle
of regolith-wayland.service and regolith-x11.service is dependent on
regolith-wayland.target and regolith-x11.target. This implies, starting
the services starts the targets and the service then responds to
lifecycle of the target, including restart and stop. But starting the
target won't start the service.

Fixes: #36